requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
python-dotenv
salesforce-merlion